requirements.txt
numpy<2.0.0
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-google-genai==1.0.3
chromadb==0.4.24
streamlit==1.32.0
python-dotenv==1.0.1
sentence-transformers==2.7.0